Moved SimulationOptions into its own file. Added some Doxygen comments.
