+++ drake/systems/Simulation.h
@@ -7,6 +7,7 @@
 #include <thread>
 
 #include "drake/core/Vector.h"
+#include "drake/systems/simulation_options.h"
 
 namespace Drake {
 
@@ -15,39 +16,6 @@ namespace Drake {
 *@brief Algorithms for simulating dynamical systems
 *@}
 */
-
-// simulation options
-struct SimulationOptions {
-  double realtime_factor;  // 1 means try to run at realtime speed, 0 is run as
-                           // fast as possible, < 0 means use default
-  double initial_step_size;
-  double timeout_seconds;
-
-  /**
-   * This variable dermines what happens if the simulation's timing is delayed
-   * by more than timeout_seconds. When this occurs, a warning is printed if
-   * this variable is true, and an exception is raised if this variable is
-   * false.
-   */
-  bool warn_real_time_violation;
-
-  /**
-   * Enables a custom simulation termination condition. This function is called
-   * each cycle of the simulation loop. The input parameter of type double is
-   * the current simulation time. If the function returns true, the simulation
-   * is terminated. The default for this is a function that always returns
-   * false.
-   */
-  std::function<bool(double)> should_stop;
-
-  SimulationOptions()
-      : realtime_factor(-1.0),
-        initial_step_size(0.01),
-        timeout_seconds(1.0),
-        warn_real_time_violation(false),
-        should_stop([](double sim_time) { return false; }) {}
-};
-static const SimulationOptions default_simulation_options;
 
 typedef std::chrono::system_clock TimeClock;  // would love to use steady_clock,
                                               // but it seems to not compile on
